文件管理系统引入

+++ yshop-pms/pms-api/src/main/java/com/fly4j/yshop/pms/pojo/entity/PmsCategory.java
@@ -9,14 +9,14 @@ import lombok.Data;
 public class PmsCategory extends BaseEntity {
   @TableId
   @JsonFormat(shape = JsonFormat.Shape.STRING)
-  private long id;
+  private Long id;
   private String name;
   private String description;
-  private long parent_id;
-  private long level;
+  private Long parent_id;
+  private Integer level;
   private String icon_url;
   private String pic_Url;
-  private long sort;
+  private Long sort;
 
 
 }
